stockplanconnect: allow ignoring journal postings with metadata key
When shares are transferred outside of Morgan Stanley rather than selling
them, StockPlanConnect does not provide any documentation that can be imported
to generate the inventory-reducing transaction.  This feature allows the user to
generate that transaction manually, and mark the posting for the
stockplanconnect-covered account to be ignored, rather than having it generate
an InvalidSourceResult.